requirementsReference.txt
absl-py==0.11.0
adal==1.2.5
aidungeon==1.0.0
alabaster==0.7.12
alembic==1.4.1
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
applicationinsights==0.11.9
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.1
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
azure-common==1.1.26
azure-core==1.10.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.7.0
azureml-automl-core==1.20.0
azureml-contrib-fairness==1.20.0
azureml-core==1.20.0
azureml-dataprep==2.7.3
azureml-dataprep-native==27.0.0
azureml-dataprep-rslex==1.5.0
azureml-dataset-runtime==1.20.0
azureml-mlflow==1.20.0.post1
azureml-pipeline==1.20.0
azureml-pipeline-core==1.20.0
azureml-pipeline-steps==1.20.0
azureml-sdk==1.20.0
azureml-telemetry==1.20.0
azureml-train==1.20.0
azureml-train-automl-client==1.20.0
azureml-train-core==1.20.0
azureml-train-restclients-hyperdrive==1.20.0
azureml-widgets==1.20.0
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
base58==2.0.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bert-tensorflow==1.0.4
bkcharts==0.2
bleach==3.1.5
blinker==1.4
blis==0.7.4
boto==2.49.0
boto3==1.16.18
botocore==1.19.18
Bottleneck==1.3.2
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.1.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.9.2
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
confuse==1.4.0
contextlib2==0.6.0.post1
cryptography==3.3.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
cytoolz==0.10.1
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
databricks-cli==0.14.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distro==1.5.0
dm-tree==0.1.5
docker==4.4.1
docutils==0.16
dotnetcore2==2.1.20
entrypoints==0.3
enum-compat==0.0.3
ezodf==0.3.2
fairlearn==0.5.0
fastcache==1.1.0
flake8==3.8.3
Flask==1.1.2
Flask-Compress==1.8.0
freeze==0.0.1
fsspec==0.7.4
fusepy==3.0.1
future==0.18.2
gast==0.3.3
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
glob2==0.7
gmpy2==2.0.8
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==0.4.16
grpcio==1.33.2
h5py==2.10.0
HeapDict==1.0.1
htmlmin==0.1.12
ImageHash==4.2.0
imagesize==1.2.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-resources==5.1.2
interpret-core==0.2.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonpickle==1.5.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
jyquickhelper==0.4.220
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.33.0+1.g022ab0f
locket==0.2.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.13.1
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msal==1.8.0
msal-extensions==0.2.2
msgpack==1.0.0
msrest==0.6.19
msrestazure==0.6.4
multipledispatch==0.6.0
murmurhash==1.0.5
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
ndg-httpsclient==0.5.1
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas-ods-reader==0.0.7
pandas-profiling==2.11.0
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathspec==0.8.1
pathtools==0.1.2
pathy==0.4.0
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
pkginfo==1.5.0.1
plotly==4.14.1
pluggy==0.13.1
ply==3.11
portalocker==1.7.1
preshed==3.0.5
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.14.0
psutil==5.7.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycurl==7.43.0.5
pydantic==1.7.3
pydeck==0.5.0
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyodbc===4.0.0-unsupported
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySide6==6.0.0
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
QtPy==1.9.0
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
rsa==4.6
Rtree==0.9.4
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-image==0.16.2
scipy==1.6.2
seaborn==0.10.1
SecretStorage==3.3.0
Send2Trash==1.5.0
shiboken6==6.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.13
six==1.15.0
smart-open==3.0.0
smmap==3.0.4
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==3.0.3
spacy-legacy==3.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
srsly==2.4.0
statsmodels==0.11.1
streamlit==0.71.0
tables==3.6.1
tabulate==0.8.7
tangled-up-in-unicode==0.0.7
tblib==1.6.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.29.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textwrap3==0.9.2
thinc==8.0.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.60.0
traitlets==4.3.3
typer==0.3.2
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.1.2
urllib3==1.26.2
validators==0.18.1
visions==0.6.0
waitress==1.4.4
wasabi==0.8.2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xgboost==1.3.3
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.3.0
xmltodict==0.12.0
yellowbrick==1.3.post1
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1